src/requirements.txt
python-dotenv==1.0.1
litellm==1.53.7